+++ include/game/component/velocity.hpp
@@ -3,11 +3,11 @@
 namespace wheel {
 
 struct VelocityComponent {
-    VelocityComponent(double max_speed) : speed(max_speed), max_speed(max_speed) {}
-    VelocityComponent(double speed, double max_speed) : speed(speed), max_speed(max_speed) {}
+    VelocityComponent(float max_speed) : speed(max_speed), max_speed(max_speed) {}
+    VelocityComponent(float speed, float max_speed) : speed(speed), max_speed(max_speed) {}
 
-    double max_speed = 0.;  // 1s pixel
-    double speed = 0.;
+    float max_speed = 0.;  // 1s pixel
+    float speed = 0.;
 };
 
 }  // namespace wheel
