+++ src/system/tower.cpp
@@ -22,13 +22,13 @@ void TowerSystem::attack() {
         : ecs.get_entity_and_components<TowerComponent, ActionsComponent, PositionComponent, AimDirectionComponent>()) {
 
         // get nearest enemy position
-        Vector2D<double> pos = {1000000, 1000000};
-        double min_distance = tower_position.vec.distance(pos);
+        Vector2D<float> pos = {1000000, 1000000};
+        float min_distance = tower_position.vec.distance(pos);
         for (auto [enemy, enemy_position] : ecs.get_components<EnemyComponent, PositionComponent>()) {
             if (!Map::is_in_bound(enemy_position.vec)) {
                 continue;
             }
-            double distance = tower_position.vec.distance(enemy_position.vec);
+            float distance = tower_position.vec.distance(enemy_position.vec);
             if (distance < min_distance) {
                 min_distance = distance;
                 pos = enemy_position.vec;
